INCOMPLETE FOR NOW
Need to make a lot of changes. For one, the code is probably very inefficient, though it is "running". Secondly, I need to add some small features to this. Thirdly, the board generator might change drastically, or be removed completely depending on feedback I receive from my professor.

Apart from that, need to create comments. A lot. Of comments.